+++ README.md
@@ -11,9 +11,9 @@ HydrideSegmentation is a comprehensive toolkit for the segmentation and analysis
 
 ### Code Structure
 
-- `hydrideseg/gui_app.py` – main Tkinter application logic.
-- `hydrideseg/analysis.py` – utilities for hydride orientation analysis and plotting.
-- `GUI.py` – entry point with optional `--debug` mode.
+- `hydride_segmentation/core/gui_app.py` – main Tkinter application logic.
+- `hydride_segmentation/core/analysis.py` – utilities for hydride orientation analysis and plotting.
+- `hydride_segmentation/gui.py` – entry point with optional `--debug` mode.
 
 ## Example Workflow
 
@@ -35,20 +35,22 @@ Below is an example output generated by the toolkit, using the test image `test_
 
 ## Getting Started
 
-1. **Install dependencies**:
+1. **Install the package locally**:
    ```bash
-   pip install -r requirements.txt
+   git submodule add <repository-url>
+   cd HydrideSegmentation
+   pip install -e .
    sudo apt-get install -y libgl1  # For OpenCV image display
    ```
 
 2. **Run the analysis**:
    ```bash
-   python hydrideOrientationAnalyzer.py --image test_data/3PB_SRT_data_generation_1817_OD_side1_8.png --shouldRunSegmentation
+   hydride-orientation --image test_data/3PB_SRT_data_generation_1817_OD_side1_8.png --shouldRunSegmentation
    ```
 
 3. **Use the GUI**:
    ```bash
-   python GUI.py
+   hydride-gui
    ```
    Running with `--debug` skips the GUI and saves a six-panel plot to
    `test_data/results/3PB_SRT_data_generation_1817_OD_side1_8_combined.png`.
@@ -66,4 +68,5 @@ If you use HydrideSegmentation in your research, please cite this repository.
 
 ---
 
-For questions or contributions, see [CONTRIBUTE.md](CONTRIBUTE.md).
+For questions or contributions, see [CONTRIBUTE.md](CONTRIBUTE.md).  
+Developers can find additional examples in [developer_guide.md](developer_guide.md).
