+++ hydride_segmentation/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core GUI and analysis modules."""
